feat: governance e2e tests added on ci and doc updates
