+++ .github/workflows/release-pipeline.yml
@@ -225,6 +225,11 @@ jobs:
         with:
           go-version: "1.25.5"
 
+      - name: Set up Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: "20"
+
       - name: Configure Git
         run: |
           git config user.name "GitHub Actions Bot"
