+++ docs/architecture/core/mcp.mdx
@@ -137,7 +137,7 @@ graph TB
 - **Security:** Similar to HTTP with streaming capabilities
 - **Benefits:** Real-time updates, persistent connections, event-driven
 
-> **MCP Configuration:** [MCP Setup Guide →](../../features/mcp)
+> **MCP Configuration:** [MCP Setup Guide →](../../mcp/overview)
 
 ---
 
@@ -272,7 +272,7 @@ curl -X POST http://localhost:8080/v1/chat/completions \
 - **Audit Trail** - Track which tools are used by which requests
 - **Risk Mitigation** - Prevent access to dangerous operations
 
-> **📖 Tool Filtering:** [MCP Tool Control →](../../features/mcp)
+> **📖 Tool Filtering:** [MCP Tool Control →](../../mcp/filtering)
 
 ---
 
@@ -918,7 +918,7 @@ mcpConfig := &schemas.MCPConfig{
 - **Failover:** Use local fallbacks when remote services are unavailable
 - **Optimization:** Route tool calls to most appropriate execution environment
 
-> **📖 MCP Development:** [Tool Development Guide →](../../features/mcp)
+> **📖 MCP Development:** [Tool Development Guide →](../../mcp/overview)
 
 ---
 
@@ -974,7 +974,7 @@ graph TB
 - **Monitoring** - Continuous security monitoring and alerting
 - **Incident Response** - Procedures for security incidents involving tools
 
-> **📖 MCP Security:** [Security Best Practices →](../../features/mcp)
+> **📖 MCP Security:** [Security Best Practices →](../../mcp/overview)
 
 ---
