+++ docs/features/governance/mcp-tools.mdx
@@ -8,7 +8,7 @@ icon: "grid-2"
 
 MCP Tool Filtering allows you to control which tools are available to AI models on a per-request basis using Virtual Keys (VKs). By configuring a VirtualKey, you can create a strict allow-list of MCP clients and tools, ensuring that only approved tools can be executed.
 
-Make sure you have at least one MCP client set up. Read more about it [here](../mcp).
+Make sure you have at least one MCP client set up. Read more about it [here](../../mcp/overview).
 
 ## How It Works
