+++ docs/mcp/agent-mode.mdx
@@ -6,7 +6,7 @@ icon: "robot"
 ---
 
 <Note>
-This feature is only available on `v1.4.0-prerelease1` and above. Please check the [changelogs](../../changelogs) for more details.
+This feature is only available on `v1.4.0-prerelease1` and above.
 </Note>
 
 ## Overview
