+++ docs/mcp/filtering.mdx
@@ -253,7 +253,7 @@ curl -X POST http://localhost:8080/api/governance/virtual-keys \
 | `tools_to_execute: ["a", "b"]` | Only specified tools |
 | Client not configured | All tools blocked from that client |
 
-Learn more in [MCP Tool Filtering for Virtual Keys](../governance/mcp-tools).
+Learn more in [MCP Tool Filtering for Virtual Keys](../features/governance/mcp-tools).
 
 ---
 
@@ -438,7 +438,7 @@ The tools included in a chat request depend on all active filters. To see what t
 ## Next Steps
 
 <CardGroup cols={2}>
-  <Card title="Virtual Key MCP Tools" icon="key" href="../governance/mcp-tools">
+  <Card title="Virtual Key MCP Tools" icon="key" href="../features/governance/mcp-tools">
     Detailed VK tool configuration
   </Card>
   <Card title="Agent Mode" icon="robot" href="./agent-mode">
