+++ docs/mcp/gateway-url.mdx
@@ -6,7 +6,7 @@ icon: "server"
 ---
 
 <Note>
-This feature is only available on `v1.4.0-prerelease1` and above. Please check the [changelogs](../../changelogs) for more details.
+This feature is only available on `v1.4.0-prerelease1` and above.
 </Note>
 
 <Info>
@@ -216,7 +216,7 @@ Configure which tools each Virtual Key can access:
 }
 ```
 
-Learn more about Virtual Key tool filtering in [MCP Tool Filtering](../governance/mcp-tools).
+Learn more about Virtual Key tool filtering in [MCP Tool Filtering](../features/governance/mcp-tools).
 
 ---
 
@@ -358,7 +358,7 @@ Consider network-level restrictions to limit which IPs can access the MCP endpoi
   <Card title="Tool Filtering" icon="filter" href="./filtering">
     Control which tools are available per request
   </Card>
-  <Card title="Virtual Key MCP Tools" icon="key" href="../governance/mcp-tools">
+  <Card title="Virtual Key MCP Tools" icon="key" href="../features/governance/mcp-tools">
     Configure per-VK tool access
   </Card>
 </CardGroup>
