+++ docs/mcp/tool-hosting.mdx
