+++ docs/quickstart/gateway/setting-up.mdx
@@ -203,9 +203,9 @@ Now that you have Bifrost running, explore these focused guides:
 ### Advanced Topics
 
 - **[Tracing](../../features/observability/default)** - Logging requests for monitoring and debugging
-- **[MCP Tools](../../features/mcp)** - Enable AI models to use external tools (filesystem, web search, databases)
-- **[Governance](../../features/governance)** - Usage tracking, rate limiting, and cost control
-- **[Deployment](../../deployment-guides)** - Production setup and scaling
+- **[MCP Tools](../../mcp/overview)** - Enable AI models to use external tools (filesystem, web search, databases)
+- **[Governance](../../features/governance/virtual-keys)** - Usage tracking, rate limiting, and cost control
+- **[Deployment](../../deployment-guides/k8s)** - Production setup and scaling
 
 ---
