+++ docs/quickstart/gateway/tool-calling.mdx
@@ -127,7 +127,7 @@ curl --location 'http://localhost:8080/api/mcp/clients'
 
 </Tabs>
 
-Read more about MCP connections and advanced end to end tool execution in the [MCP Features](../../features/mcp) section.
+Read more about MCP connections and advanced end to end tool execution in the [MCP Features](../../mcp/overview) section.
 
 ## Tool Choice Options
 
@@ -160,6 +160,6 @@ Now that you understand tool calling, explore these related topics:
 
 ### Advanced Topics
 
-- **[MCP Features](../../features/mcp)** - Advanced MCP server management and configuration
-- **[Core Features](../../features/)** - Advanced Bifrost capabilities
-- **[Architecture](../../architecture/)** - How Bifrost works internally
+- **[MCP Features](../../mcp/overview)** - Advanced MCP server management and configuration
+- **[Core Features](../../features/drop-in-replacement)** - Advanced Bifrost capabilities
+- **[Architecture](../../architecture/core/request-flow)** - How Bifrost works internally
