+++ docs/quickstart/go-sdk/tool-calling.mdx
@@ -114,7 +114,7 @@ if response.Choices[0].Message.ChatAssistantMessage != nil && response.Choices[0
 }
 ```
 
-Read more about MCP connections and in-house tool registration via local MCP server in the [MCP Features](../../features/mcp) section.
+Read more about MCP connections and in-house tool registration via local MCP server in the [MCP Features](../../mcp/overview) section.
 
 ## Advanced Tool Examples
 
@@ -265,4 +265,4 @@ Params: &schemas.ChatParameters{
 - **[Multimodal AI](./multimodal)** - Process images, audio, and multimedia content
 - **[Streaming Responses](./streaming)** - Real-time response generation
 - **[Provider Configuration](./provider-configuration)** - Multiple providers for redundancy
-- **[MCP Features](../../features/mcp)** - Advanced MCP server management
+- **[MCP Features](../../mcp/overview)** - Advanced MCP server management
