+++ tests/governance/go.mod
@@ -1,0 +1,3 @@
+module github.com/maximhq/bifrost/tests/governance
+
+go 1.25.5
